// frontend/src/lib/api/model/authUserResponseDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Student Arena API
 * API Documentation
 * OpenAPI spec version: 1.0
 */
import type { PlayerResponseDto } from './playerResponseDto';

export interface AuthUserResponseDto {
  id: number;
  email: string;
  playerAccount: PlayerResponseDto;
}